apartigas `hosting/admin.py`; iom pli strukturigas la prezenton de informeroj por admin
