+++ hosting/admin/widgets.py
@@ -1,0 +1,12 @@
+from django.contrib.admin import widgets as admin_widgets
+
+from ..widgets import ClearableWithPreviewImageInput
+
+
+class AdminImageWithPreviewWidget(ClearableWithPreviewImageInput, admin_widgets.AdminFileWidget):
+    preview_template_name = 'hosting/snippets/widget-image_file_preview_admin.html'
+
+    def format_value(self, value, **kwargs):
+        substitutions = {'initial_raw_value': value}
+        substitutions.update(**kwargs)
+        return super().format_value(value, **substitutions)
