+++ hosting/widgets.py
@@ -1,4 +1,3 @@
-from django.contrib.admin import widgets as admin_widgets
 from django.forms import widgets as form_widgets
 from django.template.loader import get_template
 from django.utils.safestring import mark_safe
@@ -27,15 +26,6 @@ class ClearableWithPreviewImageInput(form_widgets.ClearableFileInput):
         substitutions.update(**kwargs)
         rendered = mark_safe(preview_template.render(substitutions).strip())
         return self.ImagePreviewValue(value, rendered)
-
-
-class AdminImageWithPreviewWidget(ClearableWithPreviewImageInput, admin_widgets.AdminFileWidget):
-    preview_template_name = 'hosting/snippets/widget-image_file_preview_admin.html'
-
-    def format_value(self, value, **kwargs):
-        substitutions = {'initial_raw_value': value}
-        substitutions.update(**kwargs)
-        return super().format_value(value, **substitutions)
 
 
 class TextWithDatalistInput(form_widgets.TextInput):
